fix: move `/assets` to `/src/ui`
