+++ src/ui/FuzzPanelView.css
